# pecos-backendadmin-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pecos River Traders API",
        "description": "API for Pecos River Traders e-commerce platform",
        "contact": {
            "email": "admin@pecosrivertraders.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/pecos-backendadmin-api/public/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get all events",
                "description": "Get all events.",
                "operationId": "9e95997547ecbecb84ca66f1e1a1c359",
                "parameters": [
                    {
                        "name": "upcoming",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/events/upcoming": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get upcoming events",
                "description": "Get upcoming events.",
                "operationId": "00417f81e4e4f9722ee0b07627fe45ab",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event by ID",
                "description": "Get a single event.",
                "operationId": "9682ddd0801bf2c28092cf1528074303",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/admin/events": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create event (admin)",
                "description": "Admin: Create event.",
                "operationId": "af739bae60b4b370996061dca47fbb14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "EventName",
                                    "StartDate",
                                    "EndDate"
                                ],
                                "properties": {
                                    "EventName": {
                                        "type": "string"
                                    },
                                    "StartDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "EndDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "StartTime": {
                                        "type": "string"
                                    },
                                    "EndTime": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/events/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update event (admin)",
                "description": "Admin: Update event.",
                "operationId": "1c932eb68741809ed2dc37e01eced4a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event updated"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete event (admin)",
                "description": "Admin: Delete event.",
                "operationId": "5ed5c847250e8571568b0347c3156884",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event deleted"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}